server/requirements.txt
anyio==3.6.2
asgiref==3.6.0
bidict==0.22.1
certifi==2022.12.7
charset-normalizer==2.0.12
click==8.1.3
fastapi==0.75.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
loguru==0.5.3
psycopg2-binary==2.9.2
pydantic==1.8.2
PyJWT==1.7.1
PyMySQL==1.0.2
python-engineio==4.4.0
python-multipart==0.0.5
python-socketio==5.8.0
requests==2.26.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.27
starlette==0.17.1
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.17.6
